wip: improve coverage - working on delete/disabled

+++ app/src/androidTest/java/com/github/jameshnsears/quoteunquote/database/DatabaseRepositoryDouble.java
@@ -57,7 +57,7 @@ public final class DatabaseRepositoryDouble extends DatabaseRepository {
     }
 
     @NonNull
-    public Integer countReported() {
-        return reportedDAO.countReported();
+    public int countReported() {
+        return reportedDAO.countReported().intValue();
     }
 }
